Move 3DConnexion support to plugin
CURA-12407
