+++ plugins/3DConnexion/OverlayNode.py
